增加统计接口

https://codeup.aliyun.com/60e3c522f00dfa2fa4cab323/backstage/merge_request/6324940

+++ app/src/main/java/cn/atsoft/dasheng/crm/service/OrderDetailService.java
@@ -29,7 +29,7 @@ public interface OrderDetailService extends IService<OrderDetail> {
 
     OrderDetailResult record(OrderDetailParam param);
 
-    Integer addList(Long orderId, Long customerId, List<OrderDetailParam> params);
+    Long addList(Long orderId, Long customerId, List<OrderDetailParam> params);
 
     /**
      * 删除
