+++ app/src/main/java/cn/atsoft/dasheng/crm/service/impl/OrderServiceImpl.java
@@ -112,7 +112,7 @@ public class OrderServiceImpl extends ServiceImpl<OrderMapper, Order> implements
 
         param.getPaymentParam().setOrderId(entity.getOrderId());
         supplyService.OrdersBackFill(param.getSellerId(), param.getDetailParams());  //回填
-        Integer totalAmount = detailService.addList(entity.getOrderId(), param.getSellerId(), param.getDetailParams());    //返回添加所有物料总价
+        long  totalAmount = detailService.addList(entity.getOrderId(), param.getSellerId(), param.getDetailParams());    //返回添加所有物料总价
         paymentService.add(param.getPaymentParam(), orderType);
         if (ToolUtil.isNotEmpty(param.getPaymentParam()) && ToolUtil.isNotEmpty(param.getPaymentParam().getFloatingAmount())) {
             totalAmount = totalAmount + param.getPaymentParam().getFloatingAmount();
